requirements.txt
Flask>=2.1.0
Pandas
azure-functions
azure-storage-blob
azure-cosmos
azure-cosmosdb-table
Flask-Validator>=1.4.2
Flask-WTF>=1.0.1
msal>=1.18.0
networkx>=2.8.5
plotly>=5.9.0
werkzeug>2
flask-session>=0.3.2
requests>=2
msal>=1.7